requirements.txt
about-time==4.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alive-progress==3.1.5
anyio==4.3.0
appdirs==1.4.4
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
backports-datetime-fromisoformat==2.0.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.1
boto3==1.34.120
botocore==1.34.120
bs4==0.0.2
cachelib==0.13.0
certifi==2024.2.2
cffi==1.16.0
cfscrape==2.1.1
charset-normalizer==3.3.2
click==8.1.7
cloudflare-bypass==0.1.1
cloudscraper==1.2.71
colorama==0.4.6
coloredlogs==15.0.1
colorhash==2.0.0
configparser==7.0.0
construct==2.8.8
cryptography==42.0.5
Deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.19.0
exceptiongroup==1.2.1
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.6.0
Flask-Monitoring==1.1.2
Flask-OpenAPI3-UI==9.10.9
flask-profiler==1.8.1
Flask-Session==0.8.0
Flask-SocketIO==5.3.6
Flask-SSLify==0.1.5
flask-talisman==1.1.0
Flask-WTF==1.2.1
frozenlist==1.4.1
google==3.0.0
grapheme==0.6.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
limits==3.11.0
loguru==0.7.2
lxml==5.2.1
m3u8==4.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistune==3.0.2
MouseInfo==0.1.3
msgspec==0.18.6
multidict==6.0.5
numpy==1.26.4
opencv-python==4.10.0.84
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
playwright==1.45.1
prompt_toolkit==3.0.46
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pyee==11.1.0
PyExecJS==1.5.1
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.8.0
pymp4==1.4.0
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
pyppeteer==2.0.0
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-jose==3.3.0
python-socketio==5.11.2
pytweening==1.2.0
pytz==2024.1
pywidevine==1.8.0
PyYAML==6.0.1
referencing==0.35.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
scrappeycom==0.3.8
selenium==4.23.1
sentry-sdk==2.0.1
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
Sublist3r==1.0
telepot==12.7
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
Unidecode==1.3.8
urllib3==1.26.19
wcwidth==0.2.13
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.2
win32-setctime==1.1.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.19.2